# Conflicts:
#	SConstruct
#	cereal/SConscript
#	cereal/SConstruct
#	common/params.py
#	opendbc/tesla_can.dbc
#	panda/tests/pedal/enter_canloader.py
#	selfdrive/car/car_helpers.py
#	selfdrive/controls/controlsd.py
#	selfdrive/locationd/test/ubloxd.py
#	selfdrive/locationd/test/ubloxd_easy.py
#	selfdrive/locationd/ubloxd_main.cc
#	selfdrive/manager.py
#	selfdrive/modeld/SConscript
#	selfdrive/modeld/modeld.cc
#	selfdrive/test/process_replay/test_processes.py
#	selfdrive/test/test_car_models.py
#	selfdrive/thermald/thermald.py
#	selfdrive/ui/sound.cc
#	selfdrive/ui/ui.cc
#	selfdrive/ui/ui.hpp

+++ selfdrive/manager.py
@@ -7,9 +7,6 @@ import errno
 import signal
 import shutil
 import subprocess
-from selfdrive.tinklad.tinkla_interface import TinklaClient
-from cereal import tinkla
-from selfdrive.car.tesla.readconfig import CarSettings
 import datetime
 import textwrap
 from typing import Dict, List
@@ -39,7 +36,7 @@ if ANDROID:
 def unblock_stdout():
   # get a non-blocking stdout
   child_pid, child_pty = os.forkpty()
-  if child_pid != 0: # parent
+  if child_pid != 0:  # parent
 
     # child is in its own process group, manually pass kill signals
     signal.signal(signal.SIGINT, lambda signum, frame: os.kill(child_pid, signal.SIGINT))
@@ -126,12 +123,16 @@ if not prebuilt:
       compile_output += r
 
       if retry:
-        print("scons build failed, cleaning in")
-        for i in range(3, -1, -1):
-          print("....%d" % i)
-          time.sleep(1)
-        #subprocess.check_call(["scons", "-c"], cwd=BASEDIR, env=env)
-        shutil.rmtree("/tmp/scons_cache")
+        if not os.getenv("CI"):
+          print("scons build failed, cleaning in")
+          for i in range(3, -1, -1):
+            print("....%d" % i)
+            time.sleep(1)
+          subprocess.check_call(["scons", "-c"], cwd=BASEDIR, env=env)
+          shutil.rmtree("/tmp/scons_cache")
+        else:
+          print("scons build failed after retry")
+          sys.exit(1)
       else:
         # Build failed log errors
         errors = [line.decode('utf8', 'replace') for line in compile_output
@@ -142,7 +143,7 @@ if not prebuilt:
 
         # Show TextWindow
         error_s = "\n \n".join(["\n".join(textwrap.wrap(e, 65)) for e in errors])
-        with TextWindow("Openpilot failed to build\n \n" + error_s) as t:
+        with TextWindow("openpilot failed to build\n \n" + error_s) as t:
           t.wait_for_exit()
 
         exit(1)
@@ -160,20 +161,18 @@ from selfdrive.loggerd.config import ROOT
 from selfdrive.launcher import launcher
 from common import android
 from common.apk import update_apks, pm_apply_packages, start_offroad
-from common.manager_helpers import print_cpu_usage
 
 ThermalStatus = cereal.log.ThermalData.ThermalStatus
 
 # comment out anything you don't want to run
 managed_processes = {
-  "tinklad":  "selfdrive.tinklad.tinklad",
   "thermald": "selfdrive.thermald.thermald",
   "uploader": "selfdrive.loggerd.uploader",
   "deleter": "selfdrive.loggerd.deleter",
   "controlsd": "selfdrive.controls.controlsd",
   "plannerd": "selfdrive.controls.plannerd",
   "radard": "selfdrive.controls.radard",
-  "dmonitoringd": "selfdrive.controls.dmonitoringd",
+  "dmonitoringd": "selfdrive.monitoring.dmonitoringd",
   "ubloxd": ("selfdrive/locationd", ["./ubloxd"]),
   "loggerd": ("selfdrive/loggerd", ["./loggerd"]),
   "logmessaged": "selfdrive.logmessaged",
@@ -185,7 +184,7 @@ managed_processes = {
   "pandad": "selfdrive.pandad",
   "ui": ("selfdrive/ui", ["./ui"]),
   "calibrationd": "selfdrive.locationd.calibrationd",
-  "paramsd": ("selfdrive/locationd", ["./paramsd"]),
+  "paramsd": "selfdrive.locationd.paramsd",
   "camerad": ("selfdrive/camerad", ["./camerad"]),
   "sensord": ("selfdrive/sensord", ["./sensord"]),
   "clocksd": ("selfdrive/clocksd", ["./clocksd"]),
@@ -193,7 +192,7 @@ managed_processes = {
   "updated": "selfdrive.updated",
   "dmonitoringmodeld": ("selfdrive/modeld", ["./dmonitoringmodeld"]),
   "modeld": ("selfdrive/modeld", ["./modeld"]),
-  "driverview": "selfdrive.controls.lib.driverview",
+  "driverview": "selfdrive.monitoring.driverview",
 }
 
 daemon_processes = {
@@ -211,21 +210,17 @@ unkillable_processes = ['camerad']
 interrupt_processes: List[str] = []
 
 # processes to end with SIGKILL instead of SIGTERM
-kill_processes = ['sensord', 'paramsd']
+kill_processes = ['sensord']
 
 # processes to end if thermal conditions exceed Green parameters
 green_temp_processes = ['uploader']
 
 persistent_processes = [
-  'tinklad',
   'thermald',
+  'logmessaged',
   'ui',
+  'uploader',
 ]
-if not WEBCAM:
-    persistent_processes += [
-      'logmessaged',
-      'uploader',
-    ]
 
 if ANDROID:
   persistent_processes += [
@@ -238,6 +233,7 @@ if ANDROID:
 car_started_processes = [
   'controlsd',
   'plannerd',
+  'loggerd',
   'radard',
   'dmonitoringd',
   'calibrationd',
@@ -251,11 +247,7 @@ car_started_processes = [
 
 if WEBCAM:
   car_started_processes += [
-   'dmonitoringmodeld',
-  ]
-else:
-  car_started_processes += [
-    'loggerd',
+    'dmonitoringmodeld',
   ]
 
 if ANDROID:
@@ -316,11 +308,11 @@ def start_daemon_process(name):
       pass
 
   cloudlog.info("starting daemon %s" % name)
-  proc = subprocess.Popen(['python', '-m', proc],
-                         stdin=open('/dev/null', 'r'),
-                         stdout=open('/dev/null', 'w'),
-                         stderr=open('/dev/null', 'w'),
-                         preexec_fn=os.setpgrp)
+  proc = subprocess.Popen(['python', '-m', proc],  # pylint: disable=subprocess-popen-preexec-fn
+                          stdin=open('/dev/null', 'r'),
+                          stdout=open('/dev/null', 'w'),
+                          stderr=open('/dev/null', 'w'),
+                          preexec_fn=os.setpgrp)
 
   params.put(pid_param, str(proc.pid))
 
@@ -370,9 +362,11 @@ def kill_managed_process(name):
         cloudlog.critical("unkillable process %s failed to exit! rebooting in 15 if it doesn't die" % name)
         join_process(running[name], 15)
         if running[name].exitcode is None:
-          cloudlog.critical("FORCE REBOOTING PHONE!")
-          os.system("date >> /sdcard/unkillable_reboot")
-          os.system("reboot")
+          cloudlog.critical("unkillable process %s failed to die!" % name)
+          if ANDROID:
+            cloudlog.critical("FORCE REBOOTING PHONE!")
+            os.system("date >> /sdcard/unkillable_reboot")
+            os.system("reboot")
           raise RuntimeError
       else:
         cloudlog.info("killing %s with SIGKILL" % name)
@@ -404,9 +398,7 @@ def manager_init(should_register=True):
       raise Exception("server registration failed")
   else:
     dongle_id = "c"*16
-  #BB
-  if not dongle_id:
-      dongle_id = "nada"
+
   # set dongle id
   cloudlog.info("dongle id is " + dongle_id)
   os.environ['DONGLE_ID'] = dongle_id
@@ -431,38 +423,9 @@ def manager_init(should_register=True):
     os.chmod(os.path.join(BASEDIR, "cereal"), 0o755)
     os.chmod(os.path.join(BASEDIR, "cereal", "libmessaging_shared.so"), 0o755)
 
-def system(cmd):
-  try:
-    cloudlog.info("running %s" % cmd)
-    subprocess.check_output(cmd, stderr=subprocess.STDOUT, shell=True)
-  except subprocess.CalledProcessError as e:
-    cloudlog.event("running failed",
-      cmd=e.cmd,
-      output=e.output[-1024:],
-      returncode=e.returncode)
-
-def sendUserInfoToTinkla(params, tinklaClient):
-  carSettings = CarSettings()
-  gitRemote = params.get("GitRemote")
-  gitBranch = params.get("GitBranch")
-  gitHash = params.get("GitCommit")
-  dongleId = params.get("DongleId")
-  userHandle = carSettings.userHandle
-  info = tinkla.Interface.UserInfo.new_message(
-      openPilotId=dongleId,
-      userHandle=userHandle,
-      gitRemote=gitRemote,
-      gitBranch=gitBranch,
-      gitHash=gitHash
-  )
-  tinklaClient.setUserInfo(info)
-
 def manager_thread():
   # now loop
   thermal_sock = messaging.sub_sock('thermal')
-
-  if os.getenv("GET_CPU_USAGE"):
-    proc_sock = messaging.sub_sock('procLog', conflate=True)
 
   cloudlog.info("manager start")
   cloudlog.info({"environ": os.environ})
@@ -494,13 +457,6 @@ def manager_thread():
 
   logger_dead = False
 
-  # Tinkla interface
-  last_tinklad_send_attempt_time = 0
-  tinklaClient = TinklaClient()
-  sendUserInfoToTinkla(params=params, tinklaClient=tinklaClient)
-  start_t = time.time()
-  first_proc = None
-
   while 1:
     msg = messaging.recv_sock(thermal_sock, wait=True)
 
@@ -513,13 +469,6 @@ def manager_thread():
       for p in green_temp_processes:
         if p in persistent_processes:
           start_managed_process(p)
-
-    # Attempt to send pending messages if there's any that queued while offline
-    # Seems this loop runs every second or so, throttle to once every 30s
-    now = time.time()
-    if now - last_tinklad_send_attempt_time >= 30:
-      tinklaClient.attemptToSendPendingMessages()
-      last_tinklad_send_attempt_time = now
 
     if msg.thermal.freeSpace < 0.05:
       logger_dead = True
@@ -541,30 +490,14 @@ def manager_thread():
         kill_managed_process("driverview")
 
     # check the status of all processes, did any of them die?
-    # running_list = ["%s%s\u001b[0m" % ("\u001b[32m" if running[p].is_alive() else "\u001b[31m", p) for p in running]
-    #cloudlog.debug(' '.join(running_list))
+    running_list = ["%s%s\u001b[0m" % ("\u001b[32m" if running[p].is_alive() else "\u001b[31m", p) for p in running]
+    cloudlog.debug(' '.join(running_list))
 
     # Exit main loop when uninstall is needed
     if params.get("DoUninstall", encoding='utf8') == "1":
       break
 
-    if os.getenv("GET_CPU_USAGE"):
-      dt = time.time() - start_t
-
-      # Get first sample
-      if dt > 30 and first_proc is None:
-        first_proc = messaging.recv_sock(proc_sock)
-
-      # Get last sample and exit
-      if dt > 90:
-        last_proc = messaging.recv_sock(proc_sock, wait=True)
-
-        cleanup_all_processes(None, None)
-        sys.exit(print_cpu_usage(first_proc, last_proc))
-
 def manager_prepare(spinner=None):
-
-  carSettings = CarSettings()
   # build all processes
   os.chdir(os.path.dirname(os.path.abspath(__file__)))
 
@@ -573,8 +506,7 @@ def manager_prepare(spinner=None):
 
   for i, p in enumerate(managed_processes):
     if spinner is not None:
-      spinText = carSettings.spinnerText
-      spinner.update(spinText % ((100.0 - total) + total * (i + 1) / len(managed_processes),))
+      spinner.update("%d" % ((100.0 - total) + total * (i + 1) / len(managed_processes),))
     prepare_managed_process(p)
 
 def uninstall():
@@ -587,12 +519,12 @@ def uninstall():
 def main():
   os.environ['PARAMS_PATH'] = PARAMS
 
-  # the flippening!
-  if not WEBCAM:
+  if ANDROID:
+    # the flippening!
     os.system('LD_LIBRARY_PATH="" content insert --uri content://settings/system --bind name:s:user_rotation --bind value:i:1')
 
-  # disable bluetooth
-  os.system('service call bluetooth_manager 8')
+    # disable bluetooth
+    os.system('service call bluetooth_manager 8')
 
   params = Params()
   params.manager_start()
@@ -644,12 +576,9 @@ def main():
 
   try:
     manager_thread()
-  except SystemExit:
-    raise
   except Exception:
     traceback.print_exc()
     crash.capture_exception()
-    print ("EXIT ON EXCEPTION")
   finally:
     cleanup_all_processes(None, None)
